feature: provide mock module (#1478)
