+++ rm-datasource/src/test/java/io/seata/rm/datasource/mock/MockResultSetMetaData.java
@@ -1,0 +1,158 @@
+/*
+ *  Copyright 1999-2019 Seata.io Group.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package io.seata.rm.datasource.mock;
+import io.seata.rm.datasource.sql.struct.ColumnMeta;
+import java.sql.ResultSetMetaData;
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * @author will
+ * @date 2019/8/28
+ */
+public class MockResultSetMetaData implements ResultSetMetaData {
+
+    private final List<ColumnMeta> columns = new ArrayList<ColumnMeta>();
+
+    public List<ColumnMeta> getColumns() {
+        return columns;
+    }
+
+    @Override
+    public int getColumnCount() throws SQLException {
+        return columns.size();
+    }
+
+    @Override
+    public boolean isAutoIncrement(int column) throws SQLException {
+        return getColumn(column).isAutoincrement();
+    }
+
+    @Override
+    public boolean isCaseSensitive(int column) throws SQLException {
+        return false;
+    }
+
+    @Override
+    public boolean isSearchable(int column) throws SQLException {
+        return false;
+    }
+
+    @Override
+    public boolean isCurrency(int column) throws SQLException {
+        return false;
+    }
+
+    @Override
+    public int isNullable(int column) throws SQLException {
+        return 0;
+    }
+
+    @Override
+    public boolean isSigned(int column) throws SQLException {
+        return false;
+    }
+
+    @Override
+    public int getColumnDisplaySize(int column) throws SQLException {
+        return 0;
+    }
+
+    @Override
+    public String getColumnLabel(int column) throws SQLException {
+        return null;
+    }
+
+    @Override
+    public String getColumnName(int column) throws SQLException {
+        return getColumn(column).getColumnName();
+    }
+
+    @Override
+    public String getSchemaName(int column) throws SQLException {
+        return null;
+    }
+
+    @Override
+    public int getPrecision(int column) throws SQLException {
+        return 0;
+    }
+
+    @Override
+    public int getScale(int column) throws SQLException {
+        return 0;
+    }
+
+    @Override
+    public String getTableName(int column) throws SQLException {
+        return null;
+    }
+
+    @Override
+    public String getCatalogName(int column) throws SQLException {
+        return null;
+    }
+
+    @Override
+    public int getColumnType(int column) throws SQLException {
+        return 0;
+    }
+
+    @Override
+    public String getColumnTypeName(int column) throws SQLException {
+        return null;
+    }
+
+    @Override
+    public boolean isReadOnly(int column) throws SQLException {
+        return false;
+    }
+
+    @Override
+    public boolean isWritable(int column) throws SQLException {
+        return false;
+    }
+
+    @Override
+    public boolean isDefinitelyWritable(int column) throws SQLException {
+        return false;
+    }
+
+    @Override
+    public String getColumnClassName(int column) throws SQLException {
+        return null;
+    }
+
+    @Override
+    public <T> T unwrap(Class<T> iface) throws SQLException {
+        return null;
+    }
+
+    @Override
+    public boolean isWrapperFor(Class<?> iface) throws SQLException {
+        return false;
+    }
+
+    /**
+     * get column
+     * @param column
+     * @return
+     */
+    public ColumnMeta getColumn(int column){
+        return columns.get(column - 1);
+    }
+}
